Create useRegisterUser hook
